oscs via extending base PhasorTransformer
