+++ src/main/scala/buffix/ugen/Phasor.scala
@@ -1,0 +1,10 @@
+package buffix.ugen
+
+import buffix.AudioConfig
+
+/**
+  * Created by johnmcgill on 3/5/18.
+  */
+object Phasor {
+  def computeW(pitch: Int): Double = 2 * Math.PI * pitch / AudioConfig.samplingRate
+}
